+++ CHANGELOG.md
@@ -34,7 +34,7 @@ Version 1.5.1 (not released yet)
 - allow spectating in servers with Force Respawn enabled before player spawns for the first time
 - fix crash when using Switch_Model event on object with collisions disabled
 - fix replacing collision spheres when using Switch_Model event
-- calculate mesh static lighting after Auto Triggers were activated
+- recalculate mesh static lighting after Switch_Model event
 - optimize logging
 - add an option to override RF launcher by a symbolic link in the setup program